Title: 체질량지수, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240708/체질량지수/bmi.c
@@ -11,7 +11,7 @@ int main() {
 
     if(b >= 25) {
 
-        printf("%d\nObesity",(int)b);
+        printf("%.0lf\nObesity",b);
     }
